bot/requirements.txt
requests==2.31.0
beautifulsoup4==4.12.2
python-dotenv==1.0.0
schedule==1.2.1
selenium==4.21.0
openai==1.30.1
httpx==0.27.0
